requirements.txt
# ---------------------------------------------------------------------------
# Runtime dependencies for Photo-Indexer
# Install with:  pip install -r requirements.txt
# ---------------------------------------------------------------------------
# NOTE:
# * Torch CPU wheels are fetched from the official PyTorch index URL below.
# * DuckDB is optional ­– uncomment if you want columnar analytics.
# ---------------------------------------------------------------------------

--extra-index-url https://download.pytorch.org/whl/cpu

torch
torchvision
torchaudio
rawpy
piexif
Pillow
ultralytics
requests
geopy
tqdm
pydantic
click
pytest
streamlit
# Optional storage engine
# duckdb==0.10.0
